+++ src/NeoComposeUnity/Tests/NeoLoaderTests.cs
@@ -208,7 +208,7 @@ namespace NeoCompose.Tests
             var data = JsonConvert.DeserializeObject<ProjectData>(
                 @"{
   ""metadata"": {
-    ""schemaVersion"": 30,
+    ""schemaVersion"": 31,
     ""projectId"": ""project-1"",
     ""versionId"": ""version-1""
   },
@@ -594,7 +594,7 @@ namespace NeoCompose.Tests
         {
             var projectJson = $@"{{
   ""metadata"": {{
-    ""schemaVersion"": 30,
+    ""schemaVersion"": 31,
     ""projectId"": ""project-1"",
     ""versionId"": ""version-1""
   }},